% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-cmdparser.R
\name{parse_date}
\alias{parse_date}
\title{parse_date function}
\usage{
parse_date(d)
}
\arguments{
\item{d}{a date string: "YYYY-MM-DD", "YYYYMMDD", "YYYY-MM", or "YYYY". Not a vectorized function!}
}
\value{
a vector of integers: c(year, month, day), c(year, month, NA) or c(year, NA, NA)
}
\description{
parse_date function
}
\examples{
parse_date("2025-12-31")
parse_date("20251231")
parse_date("2025-12")
parse_date("2025")

}
