% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-cmdparser.R
\name{Parser}
\alias{Parser}
\title{The Parser class}
\description{
Parser class description
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{name of the application you are creating}

\item{\code{desc}}{A short description of what the software does}

\item{\code{ver}}{Software version}

\item{\code{help}}{Add -V and -h flags (default = TRUE)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Parser-new}{\code{Parser$new()}}
\item \href{#method-Parser-add_arguments}{\code{Parser$add_arguments()}}
\item \href{#method-Parser-add_commands}{\code{Parser$add_commands()}}
\item \href{#method-Parser-usage}{\code{Parser$usage()}}
\item \href{#method-Parser-parse_command_line}{\code{Parser$parse_command_line()}}
\item \href{#method-Parser-print}{\code{Parser$print()}}
\item \href{#method-Parser-clone}{\code{Parser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-new"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-new}{}}}
\subsection{Method \code{new()}}{
Initialization function (use "Parser$new")
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$new(name, desc, ver, help = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Program name}

\item{\code{desc}}{Program description string}

\item{\code{ver}}{Software version}

\item{\code{help}}{Auto-add -h (help) and -V (version) flags (default = TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-add_arguments"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-add_arguments}{}}}
\subsection{Method \code{add_arguments()}}{
Add arguments as a list of lists with entries for lparam, sparam, variable, default, type and help)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$add_arguments(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-add_commands"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-add_commands}{}}}
\subsection{Method \code{add_commands()}}{
Add commands as a list of lists with entries for command name, help and (optionally) subcmd
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$add_commands(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-usage"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-usage}{}}}
\subsection{Method \code{usage()}}{
Autogenerate a help screen
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$usage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-parse_command_line"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-parse_command_line}{}}}
\subsection{Method \code{parse_command_line()}}{
Parse the command line
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse_command_line(cmdline = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cmdline}}{Command line args}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-print"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-print}{}}}
\subsection{Method \code{print()}}{
Print function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Parser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Parser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
